Add expanded Jandex indexing OpenAPI example; refactor OpenAPI example accordingly
