+++ examples/microprofile/openapi/basic/src/main/resources/META-INF/beans.xml
