+++ examples/microprofile/openapi/basic/src/main/resources/logging.properties
