+++ examples/microprofile/openapi/basic/src/test/resources/META-INF/microprofile-config.properties
