+++ examples/microprofile/openapi/expanded-jandex/src/test/resources/META-INF/microprofile-config.properties
@@ -1,0 +1,22 @@
+#
+# Copyright (c) 2019, 2024 Oracle and/or its affiliates.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+
+# Override configuration in main source branch, so we do not use 8080 port for tests
+config_ordinal=1000
+# Microprofile server properties
+server.port=-1
+server.host=0.0.0.0
