--- examples/microprofile/openapi/src/main/java/io/helidon/microprofile/examples/openapi/GreetingProvider.java
@@ -1,48 +1,0 @@
-/*
- * Copyright (c) 2019, 2024 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.helidon.microprofile.examples.openapi;
-
-import java.util.concurrent.atomic.AtomicReference;
-
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Inject;
-import org.eclipse.microprofile.config.inject.ConfigProperty;
-
-/**
- * Provider for greeting message.
- */
-@ApplicationScoped
-public class GreetingProvider {
-    private final AtomicReference<String> message = new AtomicReference<>();
-
-    /**
-     * Create a new greeting provider, reading the message from configuration.
-     *
-     * @param message greeting to use
-     */
-    @Inject
-    public GreetingProvider(@ConfigProperty(name = "app.greeting") String message) {
-        this.message.set(message);
-    }
-
-    String getMessage() {
-        return message.get();
-    }
-
-    void setMessage(String message) {
-        this.message.set(message);
-    }
-}
